# do2425-Alex-lanero/generated-server/api/oas-doc.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Autogenerated Api
  description: Generated with OAS TOOLS CLI 

paths:
  /api/v1/flights:
    get:
      description: Returns all flights
      operationId: getFlights
      responses:
        200:
          description: Flight list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/flight'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      description: Creates a new flight
      operationId: addFlight
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/flight'
        description: Flight to be created
        required: true
      responses:
        201:
          description: Flight
        default:
            description: Unexpected error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/error'
  /api/v1/flights/{FlightNumber}:
    get:
      description: Returns a single flight by its FlightNumber
      operationId: findByFlightNumber
      parameters:
        - required: true
          name: FlightNumber
          description: FlightNumber of the flight to fetch
          in: path
          schema:
            type: number
      responses:
        200:
          description: Flight found by FlightNumber
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/flight'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    put:
      description: Update a flight based on its FlightNumber
      operationId: updateFlight
      parameters:
        - required: true
          name: FlightNumber
          description: FlightNumber of the flight to update
          in: path
          schema:
            type: number
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/flight'
        description: Flight to be updated
        required: true
      responses:
        204:
          description: Flight updated
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      description: Deletes a single flight based on its FlightNumber
      operationId: deleteFlight
      parameters:
        - required: true
          name: FlightNumber
          description: FlightNumber of the flight to delete
          in: path
          schema:
            type: number
      responses:
        204:
          description: Flight deleted
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'

components:
  schemas:
    flight:
        type: object
        required:
          - FlightNumber
        properties:
          Origin:
            type: string
          FlightNumber:
            type: number
          OnTime:
            type: boolean
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
      additionalProperties: true